Make updates for new compute studio criteria
