--- compconfig/MANIFEST.in
@@ -1,1 +1,0 @@
-include compconfig/behavior_params.json
