--- compconfig/install.sh
@@ -1,3 +1,0 @@
-# bash commands for installing your package
-BUILD_NUM=1
-conda install -c pslmodels -c conda-forge nodejs "taxbrain>=2.3.2" "paramtools>=0.7.0" pypandoc
