+++ cs-config/install.sh
@@ -1,0 +1,3 @@
+# bash commands for installing your package
+
+conda install -c pslmodels -c conda-forge nodejs "taxbrain>=2.3.2" "paramtools>=0.10.1" pypandoc
